Fix resampler
